+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * 1.1.0 (2015-??-??)
 
+ * Deprecate `IsoCodesConstraintValidator` in favor of `AbstractIsoCodesConstraintValidator`.
+
 * 1.0.3 (2015-06-02)
 
  * Add .gitattributes to ignore useless files on export.